refactor: isolate shared presentation helpers
