+++ architecture/edge-baseline.json
@@ -1,10 +1,10 @@
 {
   "app -> contracts": 7,
   "app -> legacy-contracts": 99,
-  "app -> legacy-runtime": 7,
-  "app -> node": 23,
+  "app -> legacy-runtime": 20,
+  "app -> node": 21,
   "app -> providers": 6,
-  "app -> redteam": 115,
+  "app -> redteam": 104,
   "cli -> core": 5,
   "cli -> legacy-contracts": 24,
   "cli -> legacy-runtime": 83,
@@ -24,12 +24,12 @@
   "legacy-contracts -> redteam": 11,
   "legacy-runtime -> contracts": 2,
   "legacy-runtime -> core": 12,
-  "legacy-runtime -> legacy-contracts": 68,
-  "legacy-runtime -> node": 101,
+  "legacy-runtime -> legacy-contracts": 70,
+  "legacy-runtime -> node": 103,
   "legacy-runtime -> providers": 14,
-  "legacy-runtime -> redteam": 6,
+  "legacy-runtime -> redteam": 8,
   "node -> core": 8,
-  "node -> legacy-contracts": 97,
+  "node -> legacy-contracts": 96,
   "node -> legacy-runtime": 150,
   "node -> providers": 7,
   "node -> redteam": 10,
@@ -39,8 +39,8 @@
   "providers -> node": 301,
   "providers -> redteam": 4,
   "redteam -> core": 2,
-  "redteam -> legacy-contracts": 151,
-  "redteam -> legacy-runtime": 142,
+  "redteam -> legacy-contracts": 150,
+  "redteam -> legacy-runtime": 144,
   "redteam -> node": 208,
   "redteam -> providers": 38,
   "view-server -> core": 1,
