+++ architecture/layers.json
@@ -40,6 +40,9 @@
       "allowedImportPaths": [
         "src/contracts.ts",
         "src/csv.ts",
+        "src/presentation/evalResults.ts",
+        "src/presentation/redteamConfig.ts",
+        "src/presentation/redteamMetrics.ts",
         "src/logger.ts",
         "src/providers/constants.ts",
         "src/providers/http.ts",
@@ -47,10 +50,8 @@
         "src/redteam/constants.ts",
         "src/redteam/constants/metadata.ts",
         "src/redteam/constants/strategies.ts",
-        "src/redteam/metrics.ts",
         "src/redteam/plugins/policy/utils.ts",
         "src/redteam/riskScoring.ts",
-        "src/redteam/sharedFrontend.ts",
         "src/redteam/types.ts",
         "src/tracing/store.ts",
         "src/types/api/eval.ts",
@@ -58,7 +59,6 @@
         "src/types/email.ts",
         "src/types/index.ts",
         "src/types/modelAudit.ts",
-        "src/util/convertEvalResultsToTable.ts",
         "src/util/database.ts",
         "src/util/fileExtensions.ts",
         "src/util/invariant.ts",
@@ -173,6 +173,7 @@
         "src/onboarding.ts",
         "src/openapi",
         "src/optimizer",
+        "src/presentation",
         "src/progress",
         "src/python",
         "src/remoteGrading.ts",
