+++ src/app/src/pages/eval/components/store.ts
@@ -1,5 +1,7 @@
 import { HIDDEN_METADATA_KEYS } from '@app/constants';
 import { callApi } from '@app/utils/api';
+import { convertResultsToTable } from '@promptfoo/presentation/evalResults';
+import { getRiskCategorySeverityMap } from '@promptfoo/presentation/redteamConfig';
 import { Severity } from '@promptfoo/redteam/constants';
 import {
   isPolicyMetric,
@@ -7,8 +9,6 @@ import {
   makeDefaultPolicyName,
   makeInlinePolicyId,
 } from '@promptfoo/redteam/plugins/policy/utils';
-import { getRiskCategorySeverityMap } from '@promptfoo/redteam/sharedFrontend';
-import { convertResultsToTable } from '@promptfoo/util/convertEvalResultsToTable';
 import { create } from 'zustand';
 import { persist, subscribeWithSelector } from 'zustand/middleware';
 import logger from '../../../../../logger';
