+++ src/app/src/pages/redteam/setup/utils/yamlHelpers.test.ts
@@ -12,7 +12,7 @@ vi.mock('@promptfoo/redteam/constants', () => ({
   },
 }));
 
-vi.mock('@promptfoo/redteam/sharedFrontend', () => ({
+vi.mock('@promptfoo/presentation/redteamConfig', () => ({
   getUnifiedConfig: vi.fn(),
 }));
 
@@ -20,7 +20,7 @@ vi.mock('js-yaml', () => ({
   dump: vi.fn(),
 }));
 
-import { getUnifiedConfig } from '@promptfoo/redteam/sharedFrontend';
+import { getUnifiedConfig } from '@promptfoo/presentation/redteamConfig';
 import * as yaml from 'js-yaml';
 
 describe('yamlHelpers', () => {
